move unsync OnceCell and Lazy into lazy module
